requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bids==0.0
bids-validator==1.8.9
git+https://github.com/npnl/bidsio
binaryornot==0.4.4
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
cookiecutter==1.7.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
evalutils==0.3.1
executing==0.8.2
fonttools==4.29.1
formulaic==0.2.4
idna==3.3
imageio==2.14.1
interface-meta==1.2.4
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nibabel==3.2.2
nilearn==0.9.0
notebook==6.4.8
num2words==0.5.10
numpy==1.22.1
osfclient==0.0.5
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
poyo==0.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
pybids==0.14.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.6.3
Send2Trash==1.8.0
SimpleITK==2.1.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.2.2
tomli==2.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
wget==3.2
wrapt==1.13.3
openpyxl